+++ src/Altcha.Net/AltchaService.cs
@@ -8,6 +8,7 @@ public sealed class AltchaService
     private readonly AltchaOptions _options;
     private readonly IAltchaReplayStore _replayStore;
     private readonly IAltchaClock _clock;
+    private readonly byte[] _secretKeyBytes;
 
     public AltchaService(AltchaOptions options)
         : this(options, new MemoryAltchaReplayStore())
@@ -23,6 +24,7 @@ public sealed class AltchaService
     {
         _options = options ?? throw new ArgumentNullException(nameof(options));
         _options.Validate();
+        _secretKeyBytes = Encoding.UTF8.GetBytes(_options.SecretKey);
         _replayStore = replayStore ?? throw new ArgumentNullException(nameof(replayStore));
         _clock = clock ?? throw new ArgumentNullException(nameof(clock));
     }
@@ -33,7 +35,7 @@ public sealed class AltchaService
         var salt = AltchaCrypto.RandomHex(_options.SaltLength) + "?expires=" + expires.ToString(CultureInfo.InvariantCulture) + "&";
         var number = AltchaCrypto.RandomInt(_options.Complexity.MinNumber, _options.Complexity.MaxNumber);
         var challenge = AltchaCrypto.HashHex(_options.Algorithm, salt + number.ToString(CultureInfo.InvariantCulture));
-        var signature = AltchaCrypto.HmacHex(_options.Algorithm, challenge, _options.SecretKey);
+        var signature = AltchaCrypto.HmacHex(_options.Algorithm, challenge, _secretKeyBytes);
 
         return new AltchaChallenge(_options.Algorithm, challenge, salt, signature, _options.Complexity.MaxNumber);
     }
@@ -98,7 +100,7 @@ public sealed class AltchaService
             return AltchaValidationResult.Failure(AltchaValidationError.Expired);
         }
 
-        var expectedSignature = AltchaCrypto.HmacHex(algorithm, challenge, _options.SecretKey);
+        var expectedSignature = AltchaCrypto.HmacHex(algorithm, challenge, _secretKeyBytes);
         if (!AltchaCrypto.FixedTimeEquals(expectedSignature, signature))
         {
             return AltchaValidationResult.Failure(AltchaValidationError.InvalidSignature);